+++ examples/ex02_handwritten_digits_recognition.nim
@@ -18,19 +18,19 @@ let
   mnist = load_mnist(cache = true)
   # Training data is 60k 28x28 greyscale images from 0-255,
   # neural net prefers input rescaled to [0, 1] or [-1, 1]
-  x_train = mnist.train_images.astype(float32) / 255'f32
+  x_train = mnist.train_images.asType(float32) / 255'f32
 
   # Change shape from [N, H, W] to [N, C, H, W], with C = 1 (unsqueeze). Convolution expect 4d tensors
   # And store in the context to track operations applied and build a NN graph
   X_train = ctx.variable x_train.unsqueeze(1)
 
   # Labels are uint8, we must convert them to int
-  y_train = mnist.train_labels.astype(int)
+  y_train = mnist.train_labels.asType(int)
 
   # Idem for testing data (10000 images)
-  x_test = mnist.test_images.astype(float32) / 255'f32
+  x_test = mnist.test_images.asType(float32) / 255'f32
   X_test = ctx.variable x_test.unsqueeze(1)
-  y_test = mnist.test_labels.astype(int)
+  y_test = mnist.test_labels.asType(int)
 
 # Configuration of the neural network
 network DemoNet:
